Refuse a package that promises more Reference Audio clips than we read
`max_reference_count` is published verbatim to callers as
`synth_voice_profile_capabilities_t::max_reference_count` -- a promise about
how many Reference Audio clips a Voice Profile may be built from. Both
families' Voice Profile creation reads `references[0]` and nothing else, and
both refuse `reference_count != 1` by name (src/voice-profile.cpp, documented
at both sites). Nothing constrained the published value to agree, so a
package declaring 4 advertised 4 and then had every 2-clip call refused.

Both `read_profile_contract`s now require exactly 1 rather than merely
non-zero. The rule is the same for both families because what they implement
is the same: OmniVoice reads one clip, Qwen3-TTS reads one clip. Fixed at
load, where a wrong package can still be re-cut, and NOT by weakening the
`!= 1` creation gate -- that gate names the limitation in the code itself
rather than leaning on every future package agreeing with it, which is
exactly why it is there. Both gates now carry a line saying the load rule
backs them and why they stay anyway.

Published packages are unaffected. `scripts/convert-omnivoice.py` writes the
literal 1; `scripts/convert-qwen3-tts.py` writes the package manifest's own
value, and the published Base package's committed Golden Manifest declares 1.
Verified against the real artifact rather than the script: reading
`synthesize.reference.max_reference_count` out of
`qwen3-tts-12hz-0-6b-base-BF16.gguf` gives 1, and driving that file through
`read_hparams` under the new rule returns SYNTH_OK. CustomVoice packages
carry no profile contract at all, so this function never runs for them.

The Golden Manifest schema moves with the rule. It said `"minimum": 1`, which
validated a manifest declaring 2 -- schema-valid, converts without complaint,
and produces a GGUF the loader now refuses at load. Under this project's
"docs are confirmed contracts" convention that disagreement is itself the
defect, so `max_reference_count` is `"const": 1`, with the reason and the
condition for widening it recorded in the property's own description.

The same change removes the narrowing underneath it. The field is uint64_t in
the package, in ProfileContract and in the public struct (which was ALREADY
uint64_t -- no ABI change here, and nothing under include/ is touched), but
VoiceProfileInfo held it as uint32_t -- one lossy hop in an otherwise 64-bit
path, on which a declared 2^32 became 0, the value docs/c-interface.md reads
as "no Reference Audio clip may be used at all". A package that had just been
accepted for declaring a non-zero count would then advertise none. With the
field widened, voice-profile.cpp's widening local for the ABI write is gone
too: it now writes the field directly, like every neighbouring call.

Tests: each family's load-rejection matrix gains a 2 case and a 2^32 case
(the zero case stays). Both were checked by reverting the rule to `== 0`:
both families' metadata tests then FAIL, and pass again with it restored. The
2^32 case takes the same `!= 1` branch as the 2 -- it is a regression value,
not a second mechanism, and its label and comment now say only that. The type
change is pinned separately by
test_a_declared_reference_count_is_published_unnarrowed, which drives
fill_voice_profile_capability directly -- there is no load rule in front of
that function, which is the whole reason it is unit-testable -- and asserts 2,
2^32 and UINT64_MAX all arrive intact. Reverting the field to uint32_t makes
that test FAIL. The schema change is pinned by a new case in
test_golden_manifests.py that mutates only that field on a real committed
manifest, asserting the unmutated copy still validates; relaxing the schema
back to `"minimum": 1` makes it FAIL.

+++ src/model-info.h
@@ -124,7 +124,18 @@ struct VoiceProfileInfo {
     uint64_t            min_frames_per_clip          = 0;
     uint64_t            max_frames_per_clip          = 0;
     uint64_t            max_total_frames             = 0;
-    uint32_t            max_reference_count          = 0;
+    // uint64_t, matching both ends it sits between: the package's own
+    // `synthesize.reference.max_reference_count` (a GGUF u64, so a
+    // ProfileContract u64) and the public
+    // `synth_voice_profile_capabilities_t::max_reference_count`. It was
+    // uint32_t until 2026-08-13, which made this the one narrowing hop in an
+    // otherwise 64-bit path -- a declared 2^32 arrived here as 0, the value
+    // that means "no Reference Audio clip may be used at all". Both families
+    // now refuse anything but 1 at load (weights.cpp's
+    // read_profile_contract), so nothing reaches this field that a uint32_t
+    // could not have held either; the type simply stops being the place a
+    // future widening of that rule would have to remember to look.
+    uint64_t            max_reference_count          = 0;
     uint8_t             compatibility_id[32]         = {};  // decoded from the package hex
     // The package's own declared Serialized Profile schema identity (the
     // same "synthesize.profile.schema"/"schema_version" ProfileContract pair
